Relocate now-used-once checkPayload helper
